Combine imports and exports into one transfer table with asset register
I had a table for import requests and a separate one for export requests,
which were linked with tables for imported assets and exported assets.
But these are combined now into a transfer table, with a import/export flag.
Similarly, the assets have been combined into one asset register.
However, I think some work is still needed to enable us to track all activity
for each asset. I.e. when was this asset imported, when was it exported?
Need to think about this more.