Added new metavariables to DSA schema
It's helpful to capture data about DSAs to aid reporting later on.
So I've added more columns to the DSAs schema to capture extra
information. For example, splitting out the DSA expiry date and
data destruction date, to allow for data retention periods. Also
recording a primary contact  for the data owner.