Refactored code away from DSAs form
To enable proper testing of the code, it's good practice to
keep the code actually doing things outside of the Windows forms.
That way tests can be run without instantiating a form. Code that
remains in the form is purely for working with the form objects
and values, passing their values for refactored methods, or coding
in actions of form buttons.