Refactor FillDataOwnerGridView to move code away from Form
The Add Data Owner form still contains loads of code which belongs
in a separate class. Keeping code away from the form is good practice
because, among other things, it allows for testing of the code without
having to run through the GUI. Previously a method to fill a data grid
view in the form contain a LINQ query to create the view. The LINQ
query has been moved into DSA static class, in a method called
CreateDataOwnerGridView.